% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_realtest.R
\docType{package}
\name{about_realtest}
\alias{-package}
\alias{about_realtest}
\alias{realtest}
\alias{realtest-package}
\title{Where Expectations Meet Reality: Realistic Unit Testing in R}
\description{
\pkg{realtest} is a framework for unit testing for
realistic minimalists, where we distinguish between
expected, acceptable, and undesirable behaviour.

\bold{Keywords}: unit testing, software quality, expectation,
undesired behaviour, continuous integration.

\bold{License}: GNU General Public License version 2 or later.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
