% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{ignore_differences}
\alias{ignore_differences}
\alias{identical_or_TRUE}
\alias{maps_comparer}
\alias{maps_identical_or_TRUE}
\title{Object and Side Effects Comparers}
\usage{
ignore_differences(x, y)

identical_or_TRUE(x, y)

maps_comparer(x, y, element_comparer)

maps_identical_or_TRUE(x, y)
}
\description{
return TRUE or FALSE or error message....
\code{\link[base]{identical}}
\code{\link[base]{all.equal}}
\code{identical_or_TRUE}....
\code{maps_identical_or_TRUE}...
\code{maps_comparer} compares two named lists as sets of key-value pairs (order does not matter)
treats NULLs as empty lists
does not distinguish between NULL fields and missing fields
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
