% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{...}
\usage{
E(...)
}
\arguments{
\item{...}{}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
# ...

}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
