% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R}
\alias{R}
\title{Create a Test Descriptor by Recording Effects of a Call}
\usage{
R(expr)
}
\arguments{
\item{expr}{a formal argument to be evaluated}
}
\value{
A list of class \code{realtest_descriptor},
see \code{\link{P}}, which this function calls.
Additional named component \code{expr} gives the
expression used to generate the \code{value}.
}
\description{
Evaluates an expression and records its result as well as
whether it generates any errors, warnings, or messages
and if it prints anything on
\code{\link[base]{stdout}} or \code{\link[base]{stderr}}.
}
\details{
There may be other side effects, like changing the state of
a random number generator or plotting, but these are not captured,
at least, not by the current version of the package.
}
\examples{
R(sum(1:10))
R(cat("a bit talkative, innit?"))
R(sqrt(c(-1, 0, 1, 2, 4)))
R(log("aaaargh"))

}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}

\code{\link{E}}, \code{\link{P}}
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
