% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{ignore_differences}
\alias{ignore_differences}
\alias{identical_or_TRUE}
\alias{maps_comparer}
\alias{maps_identical_or_TRUE}
\title{Example Object and Side Effect Comparers}
\usage{
ignore_differences(x, y)

identical_or_TRUE(x, y)

maps_comparer(x, y, element_comparer)

maps_identical_or_TRUE(x, y)
}
\arguments{
\item{x}{TODO}

\item{y}{TODO}

\item{element_comparer}{TODO}
}
\value{
TRUE or FALSE or error message....
\code{\link[base]{identical}}
\code{\link[base]{all.equal}}
write a custom one if you want attributes to be omitted
\code{identical_or_TRUE}....
\code{maps_identical_or_TRUE}...
\code{maps_comparer} compares two named lists as sets of key-value pairs (order does not matter)
treats NULLs as empty lists
does not distinguish between NULL fields and missing fields
}
\description{
Two-argument functions to compare direct or indirect effects
of two test descriptors (see \code{\link{P}} and \code{\link{R}}).
These can be passed as \code{value_comparer} and \code{sides_comparer}
to \code{\link{E}}.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
