% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessors.R
\name{stop_if_results_different}
\alias{stop_if_results_different}
\title{Fail on Unmet Expectations}
\usage{
stop_if_results_different(r)
}
\arguments{
\item{r}{object of class \code{realtest_result}, see \code{\link{E}}}
}
\value{
Returns \code{r}, invisibly.
}
\description{
This is an example result postprocessor that can be used in \code{\link{E}}.
Calls \code{\link[utils]{str}(r)} and throws an error
if an expectation is not met, i.e., when \code{r[["matches"]]}
is of length 0.
}
\details{
You can always create a function \code{str.realtest_result}
implementing the pretty printing of an error message.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
