% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{test_dir}
\alias{test_dir}
\title{Gather All Test Results From R Scripts}
\usage{
test_dir(
  path = "tests",
  pattern = "^realtest-.*\\\\.R",
  recursive = FALSE,
  ignore.case = FALSE
)
}
\arguments{
\item{path}{directory with scripts to execute}

\item{pattern}{regular expression specifying the file names to execute}

\item{recursive}{logical, see \code{\link[base]{list.files}}}

\item{ignore.case}{logical, see \code{\link[base]{list.files}}}
}
\value{
List of all test results, see \code{\link{E}}
}
\description{
Executes all R scripts in a given directory whose names match a given
pattern and gathers all test result in a single list,
which you can process however you desire.
}
\author{
\href{https://www.gagolewski.com/}{Marek Gagolewski}
}
\seealso{
The official online manual of \pkg{realtest} at \url{https://realtest.gagolewski.com/}
}
